refactor: move common outside features
