+++ src/features/auth/auth.controller.ts
@@ -2,7 +2,7 @@ import { RouterContext } from '@koa/router';
 import { User } from '@/models/user.model';
 import { ISignInDto } from '@/interfaces/dtos/auth/sign-in.dto';
 import { ISignUpDto } from '@/interfaces/dtos/auth/sign-up.dto';
-import { generateToken } from '@/utils/jwt/generate-token';
+import { generateToken } from '@/features/auth/utils/generate-token';
 import { addSendEmailJob } from '@/jobs/send-email.job';
 import { environment } from '@/config/environment';
