+++ src/features/auth/utils/generate-token.ts
