+++ src/features/users/users.router.ts
@@ -1,12 +1,12 @@
 import Router from '@koa/router';
 import { object, string } from 'yup';
 import { UsersController } from '@/features/users/users.controller';
-import { vaidateRequestBodyMiddleware } from '@/features/common/middlewares/validation/validate-request-body.middleware';
-import { vaidateRequestParamsMiddleware } from '@/features/common/middlewares/validation/validate-request-params.middleware';
+import { vaidateRequestBodyMiddleware } from '@/common/middlewares/validation/validate-request-body.middleware';
+import { vaidateRequestParamsMiddleware } from '@/common/middlewares/validation/validate-request-params.middleware';
 import { findUserByIdMiddleware } from '@/features/users/middlewares/find-user-by-id.middleware';
-import { isCurrentUserMiddleware } from '@/features/common/middlewares/authorization/is-current-user.middleware';
+import { isCurrentUserMiddleware } from '@/common/middlewares/authorization/is-current-user.middleware';
 import { IUpdateUserDto } from '@/interfaces/dtos/users/update-user.dto';
-import { objectId } from '@/features/common/yup/custom-schemas/object-id.schema';
+import { objectId } from '@/common/yup/custom-schemas/object-id.schema';
 
 export const usersRouter = new Router({ prefix: '/users' });
